chore: split process-logs into modules
